🛠[회원가입_가입완료] : 가입완료 컨테이너 생성
